requirements.txt
argon2-cffi==20.1.0
# asgiref==3.4.1
attrs==21.2.0
backcall==0.2.0
backports.entry-points-selectable==1.1.0
beautifulsoup4==4.9.3
# bleach==4.0.0
bs4==0.0.1
certifi==2021.5.30
cffi==1.14.6
# charset-normalizer==2.0.4
# click==8.0.1
colorama==0.4.4
cycler==0.10.0
debugpy==1.4.1
# decorator==5.0.9
defusedxml==0.7.1
distlib==0.3.2
Django==3.2.6
djangorestframework==3.12.4
entrypoints==0.3
filelock==3.0.12
idna==3.2
ipykernel==6.2.0
ipython==7.26.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.2
mistune==0.8.4
mysqlclient==2.0.3
nbclient==0.5.4
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
nltk==3.6.2
notebook==6.4.3
numpy==1.21.2
packaging==21.0
pandas==1.3.2
pandocfilters==1.4.3
parso==0.8.2
phonenumbers==8.12.29
pickleshare==0.7.5
Pillow==8.3.1
platformdirs==2.2.0
prometheus-client==0.11.0
prompt-toolkit==3.0.19
pycparser==2.20
Pygments==2.10.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.1
pywin32==301
pywinpty==1.1.3
pyzmq==22.2.1
qtconsole==5.1.1
QtPy==1.9.0
regex==2021.8.28
requests==2.26.0
scipy==1.7.1
Send2Trash==1.8.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
terminado==0.11.0
testpath==0.5.0
tornado==6.1
tqdm==4.62.2
traitlets==5.0.5
urllib3==1.26.6
virtualenv==20.7.2
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==5.3.0
widgetsnbextension==3.5.1
